ratings-backend/requirements.txt
asgiref==3.4.1
channels==3.0.4
channels-redis==3.3.1
daphne==3.0.2
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1